refactor(midnight-js): move the contracts utils under src/internal
src/utils held a barrel that re-exported three modules, none of which the
package publishes: the exports map has a single "." subpath, so utils was
never a build entry, and nothing in the index export graph re-exported a
name from it. The code is not dead -- it is imported and executed on
every call and deploy path -- but its names are unreachable to consumers,
which is what src/internal already means here. ADR 0004 records that
contracts/src/internal marks implementation hidden from consumers.

The package was using three mechanisms for one intent: src/internal, an
unexported sibling directory, and the "intentionally NOT re-exported"
comment in governance/index.ts. This removes the middle one.

The directory moves whole, so its sibling imports are unaffected. Ten
importers are repointed, two of which matter more than the rest:

- vi.mock('../utils') in unproven-call-tx.test.ts and
  unproven-deploy-tx.test.ts. A stale mock path stops mocking silently;
  left unfixed it takes six tests into real WASM deserialization.
- ledger8-entry.ts and ledger8-pipeline.ts reach the modules from inside
  src/internal, so their specifiers shorten rather than lengthen.

Seven path references in live docs and comments follow the move. The
v4.1.0 release notes keep the old path: they describe the layout of that
release, not of the tree.
