+++ packages/contracts/src/internal/utils/ledger-utils.ts
