+++ packages/protocol/src/lib/v8/execute.ts
@@ -156,7 +156,7 @@ export interface Ledger8ExecutionRuntime {
  * `zswapLocalState` is the post-call Zswap local state, DECODED into the
  * runtime's public shape: the coins the circuit spent and produced. A caller
  * turns it into the transaction's segmented Zswap offer
- * (`zswapStateToSegmentedOffer`, `packages/contracts/src/utils/zswap-utils.ts`)
+ * (`zswapStateToSegmentedOffer`, `packages/contracts/src/internal/utils/zswap-utils.ts`)
  * and hands that offer to whichever composition leg it targets.
  *
  * @see {@link RetainedEraExecution}
